+++ app/src/androidTest/kotlin/com/yapt/planttracker/ui/screens/plantlist/PlantListScreenTest.kt
@@ -59,7 +59,6 @@ class PlantListScreenTest {
             PlantListScreen(
                 viewModel = viewModel,
                 onNavigateToPlant = {},
-                onNavigateToPlantWithSuggestion = { _, _ -> },
                 onNavigateToAdd = {},
                 onNavigateToSettings = {}
             )
@@ -79,7 +78,6 @@ class PlantListScreenTest {
             PlantListScreen(
                 viewModel = viewModel,
                 onNavigateToPlant = {},
-                onNavigateToPlantWithSuggestion = { _, _ -> },
                 onNavigateToAdd = {},
                 onNavigateToSettings = {}
             )